pyproject.toml
[tool.poetry]
name = "ddua-embeddings"
version = "0.1.0"
description = ""
authors = ["Andrea D'Agostino <andrea@dagtech.io>"]
readme = "README.md"
packages = [ { include = "src", from = "." } ]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
streamlit = "^1.28.2"
langchain = "^0.0.339"
openai = "^1.3.5"
trafilatura = "^1.6.2"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
ruff = "^0.1.6"
chromadb = "^0.4.18"
tiktoken = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
